requirements.txt
claude-agent-sdk>=0.1.8
python-dotenv>=1.0.0
anyio>=4.0.0
